+++ tg_bot/__init__.py
@@ -1,1 +1,1 @@
-from .handlers import register_all_handlers
+from .handlers import register_admin_handlers, register_client_handlers, register_start_handlers
