+++ tg_bot/handlers/__init__.py
@@ -1,1 +1,3 @@
-from .main import register_all_handlers
+from .main import register_start_handlers
+from .client import register_client_handlers
+from .admin import register_admin_handlers
